chore: 기본 패키지 및 폴더 구조 세팅
