+++ build.gradle
@@ -37,6 +37,7 @@ dependencies {
     implementation 'mysql:mysql-connector-java:8.0.33'
 
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.6'
+    implementation 'org.springframework.boot:spring-boot-starter-webflux'
 
 }
