--- src/main/java/com/samsungnomads/wheretogo/docs/SwaggerConfig.java
@@ -1,21 +1,0 @@
-package com.samsungnomads.wheretogo.docs;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(apiInfo());
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .title("wheretogo API");
-    }
-}
