+++ src/main/java/com/samsungnomads/wheretogo/domain/filter/controller/FilterController.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.filter.controller;
+
+public class FilterController {
+}
