+++ src/main/java/com/samsungnomads/wheretogo/domain/filter/dto/FilterRequestDto.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.filter.dto;
+
+public class FilterRequestDto {
+}
