+++ src/main/java/com/samsungnomads/wheretogo/domain/filter/dto/FilterResponseDto.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.filter.dto;
+
+public class FilterResponseDto {
+}
