+++ src/main/java/com/samsungnomads/wheretogo/domain/filter/entity/Filter.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.filter.entity;
+
+public class Filter {
+}
