+++ src/main/java/com/samsungnomads/wheretogo/domain/filter/entity/FilterStation.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.filter.entity;
+
+public class FilterStation {
+}
