+++ src/main/java/com/samsungnomads/wheretogo/domain/filter/repository/FilterRepository.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.filter.repository;
+
+public class FilterRepository {
+}
