+++ src/main/java/com/samsungnomads/wheretogo/domain/filter/repository/FilterServiceRepository.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.filter.repository;
+
+public class FilterServiceRepository {
+}
