+++ src/main/java/com/samsungnomads/wheretogo/domain/filter/service/FilterService.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.filter.service;
+
+public class FilterService {
+}
