+++ src/main/java/com/samsungnomads/wheretogo/domain/member/controller/MemberController.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.member.controller;
+
+public class MemberController {
+}
