+++ src/main/java/com/samsungnomads/wheretogo/domain/member/dto/MemberRequestDto.java
@@ -1,0 +1,5 @@
+package com.samsungnomads.wheretogo.domain.member.dto;
+
+public class MemberRequestDto {
+    
+}
