+++ src/main/java/com/samsungnomads/wheretogo/domain/member/dto/MemberResponseDto.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.member.dto;
+
+public class MemberResponseDto {
+}
