+++ src/main/java/com/samsungnomads/wheretogo/domain/member/entity/Member.java
@@ -1,0 +1,19 @@
+package com.samsungnomads.wheretogo.domain.member.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+public class Member {
+
+    @Id
+    @GeneratedValue
+    private String id;
+    private String name;
+
+}
