+++ src/main/java/com/samsungnomads/wheretogo/domain/member/entity/MemberFilter.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.member.entity;
+
+public class MemberFilter {
+}
