+++ src/main/java/com/samsungnomads/wheretogo/domain/member/repository/MemberFilterRepository.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.member.repository;
+
+public class MemberFilterRepository {
+}
