+++ src/main/java/com/samsungnomads/wheretogo/domain/member/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.member.service;
+
+public class MemberService {
+}
