+++ src/main/java/com/samsungnomads/wheretogo/domain/station/controller/StationController.java
@@ -1,0 +1,12 @@
+package com.samsungnomads.wheretogo.domain.station.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("api/station")
+@RequiredArgsConstructor
+public class StationController {
+
+}
