+++ src/main/java/com/samsungnomads/wheretogo/domain/station/dto/StationRequestDto.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.station.dto;
+
+public class StationRequestDto {
+}
