+++ src/main/java/com/samsungnomads/wheretogo/domain/station/dto/StationResponseDto.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.station.dto;
+
+public class StationResponseDto {
+}
