+++ src/main/java/com/samsungnomads/wheretogo/domain/station/entity/Station.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.station.entity;
+
+public class Station {
+}
