+++ src/main/java/com/samsungnomads/wheretogo/domain/station/repository/StationRepository.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.station.repository;
+
+public class StationRepository {
+}
