+++ src/main/java/com/samsungnomads/wheretogo/domain/station/service/StationService.java
@@ -1,0 +1,4 @@
+package com.samsungnomads.wheretogo.domain.station.service;
+
+public class StationService {
+}
