--- src/main/java/com/samsungnomads/wheretogo/exception/GlobalExceptionHandler.java
@@ -1,100 +1,0 @@
-package com.samsungnomads.wheretogo.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindException;
-import org.springframework.web.HttpRequestMethodNotSupportedException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
-
-import java.nio.file.AccessDeniedException;
-
-/**
- * 글로벌 예외 처리 핸들러
- * 🚨 애플리케이션 전체에서 발생하는 예외를 처리
- */
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    /**
-     * 비즈니스 예외 처리
-     * 📊 서비스 계층에서 발생하는 비즈니스 예외 처리
-     */
-    @ExceptionHandler(BusinessException.class)
-    protected ResponseEntity<ErrorResponse> handleBusinessException(final BusinessException e) {
-        log.error("비즈니스 예외: {}", e.getMessage());
-        final ErrorCode errorCode = e.getErrorCode();
-        final ErrorResponse response = ErrorResponse.of(errorCode);
-        return new ResponseEntity<>(response, errorCode.getStatus());
-    }
-    
-    /**
-     * @Valid 유효성 검사 예외 처리
-     * 🧩 요청 바디 유효성 검사 실패 처리
-     */
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    protected ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(final MethodArgumentNotValidException e) {
-        log.error("유효성 검사 예외: {}", e.getMessage());
-        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_INPUT_VALUE, e.getBindingResult());
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-    
-    /**
-     * @ModelAttribute 예외 처리
-     * 🔄 모델 속성 바인딩 실패 처리
-     */
-    @ExceptionHandler(BindException.class)
-    protected ResponseEntity<ErrorResponse> handleBindException(final BindException e) {
-        log.error("바인딩 예외: {}", e.getMessage());
-        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_INPUT_VALUE, e.getBindingResult());
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-    
-    /**
-     * 지원하지 않는 HTTP 메소드 예외
-     * 🚫 지원하지 않는 HTTP 메소드 호출 처리
-     */
-    @ExceptionHandler(HttpRequestMethodNotSupportedException.class)
-    protected ResponseEntity<ErrorResponse> handleHttpRequestMethodNotSupportedException(final HttpRequestMethodNotSupportedException e) {
-        log.error("지원하지 않는 HTTP 메소드: {}", e.getMessage());
-        final ErrorResponse response = ErrorResponse.of(ErrorCode.METHOD_NOT_ALLOWED);
-        return new ResponseEntity<>(response, HttpStatus.METHOD_NOT_ALLOWED);
-    }
-    
-    /**
-     * 인수 타입 불일치 예외
-     * 🔢 요청 파라미터 타입 불일치 처리
-     */
-    @ExceptionHandler(MethodArgumentTypeMismatchException.class)
-    protected ResponseEntity<ErrorResponse> handleMethodArgumentTypeMismatchException(final MethodArgumentTypeMismatchException e) {
-        log.error("타입 불일치 예외: {}", e.getMessage());
-        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_INPUT_VALUE, e);
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-    
-    /**
-     * 접근 거부 예외
-     * 🔒 인증된 사용자가 권한이 없는 리소스 접근 시 처리
-     */
-    @ExceptionHandler(AccessDeniedException.class)
-    protected ResponseEntity<ErrorResponse> handleAccessDeniedException(final AccessDeniedException e) {
-        log.error("접근 거부 예외: {}", e.getMessage());
-        final ErrorResponse response = ErrorResponse.of(ErrorCode.INTERNAL_SERVER_ERROR);
-        return new ResponseEntity<>(response, HttpStatus.FORBIDDEN);
-    }
-    
-    /**
-     * 그 외 모든 예외 처리
-     * 💥 처리되지 않은 모든 예외를 잡아 내부 서버 오류로 처리
-     */
-    @ExceptionHandler(Exception.class)
-    protected ResponseEntity<ErrorResponse> handleException(final Exception e) {
-        log.error("처리되지 않은 예외: ", e);
-        final ErrorResponse response = ErrorResponse.of(ErrorCode.INTERNAL_SERVER_ERROR);
-        return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
