--- src/main/java/com/samsungnomads/wheretogo/member/Member.java
@@ -1,17 +1,0 @@
-package com.samsungnomads.wheretogo.member;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter @Setter
-public class Member {
-
-    @Id @GeneratedValue
-    private String id;
-    private String name;
-
-}
