--- src/main/java/com/samsungnomads/wheretogo/member/MemberRepository.java
@@ -1,18 +1,0 @@
-package com.samsungnomads.wheretogo.member;
-
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class MemberRepository {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    public String save(Member member) {
-        em.persist(member);
-        return member.getId();
-    }
-
-}
