--- src/main/java/com/samsungnomads/wheretogo/member/MemberService.java
@@ -1,4 +1,0 @@
-package com.samsungnomads.wheretogo.member;
-
-public class MemberService {
-}
