--- src/main/java/com/samsungnomads/wheretogo/member/dto/MemberRequestDto.java
@@ -1,7 +1,0 @@
-package com.samsungnomads.wheretogo.member.dto;
-
-public class MemberRequestDto {
-
-
-
-}
