--- src/main/java/com/samsungnomads/wheretogo/member/dto/MemberResponseDto.java
@@ -1,4 +1,0 @@
-package com.samsungnomads.wheretogo.member.dto;
-
-public class MemberResponseDto {
-}
