+++ src/test/java/com/samsungnomads/wheretogo/exception/TestExceptionController.java
@@ -1,5 +1,8 @@
 package com.samsungnomads.wheretogo.exception;
 
+import com.samsungnomads.wheretogo.global.error.ErrorCode;
+import com.samsungnomads.wheretogo.global.error.exception.BusinessException;
+import com.samsungnomads.wheretogo.global.error.exception.EntityNotFoundException;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/test")
 public class TestExceptionController {
-    
+
     /**
      * 비즈니스 예외 발생 엔드포인트
      * 💼 INVALID_INPUT_VALUE 예외 발생
@@ -21,7 +24,7 @@ public class TestExceptionController {
     public String throwBusinessException() {
         throw new BusinessException(ErrorCode.INVALID_INPUT_VALUE);
     }
-    
+
     /**
      * 엔티티 조회 실패 예외 발생 엔드포인트
      * 👤 회원 조회 실패 예외 발생
@@ -30,7 +33,7 @@ public class TestExceptionController {
     public String throwEntityNotFoundException(@PathVariable String id) {
         throw EntityNotFoundException.memberNotFound(id);
     }
-    
+
     /**
      * 처리되지 않은 예외 발생 엔드포인트
      * 💥 런타임 예외 발생
